[microrss] Integrate WIMM SDK for the layouts. We have used AdapterViewTray in the same view that the list layouts work, it handles the swipe left/right. Also we inherit the activities from LauncherActivity to handle the swipe up for exiting the application. We have used this too to mock the swipe up from the story activity, until we integrate another component more flexible for swipe up/down/left/right.
It removes the NavigationProcessor: to pass the ids, position for just two activities that needs specific id's seems a bit overhead. Apart from that, we needed another SQL when you try to navigate between activities. It seems far easier now, we just pass the feed id (in the future maybe story id) between activities, no matter in which position of any in-memory array.

We have created a StoryView to fix some scroll down issues, seems unbeliavable but the same View (ScrollableView) behaves differently if you put it on the XML than a custom class.

We don't pass anymore nothing from the InitActivity, it seems better to handle the different errors (no active feeds, etc.) on the correspondient activities, so that we already have the swipe left/right functionality.

Code review more than welcome, lots of changes.

re #253 fixes #254