[microrss] Integrate WIMM Network API. Now we always ask for a connection before using it. Remove some information logs, and add new ones to understand what's going on.
Now, on initActivity we only start the sync if there are no feeds. Then, the ync will be managed by the WIMM sync manager or our alarm manager (we have a flag for saying WIMM or mobile phone).

On the WIMM emulator doesn't work, even their sample for network. So, we need to test directly on the device (I am making the commit without testing on the device, but I am quite confident).

I have modified some internal things of how the service behaves and so on, so a BIG REVIEW is needed.

We still need to do a service for google reader, in the same way we have done ours.